refactored page.tsx to handle multiple routes - used react-router-dom for routing - created AppRouter.tsx component to handle routing - configured next.config.ts to include AppRouter paths
