+++ frontend/neuromance/next.config.ts
@@ -1,7 +1,14 @@
 import type { NextConfig } from "next";
 
 const nextConfig: NextConfig = {
-  /* config options here */
+  async rewrites() {
+    return [
+      {
+        source: "/:path*",
+        destination: "/", // All unmatched paths go to the root page
+      },
+    ];
+  },
 };
 
 export default nextConfig;
