+++ frontend/neuromance/src/app/components/ReadMuse/ReadMuse.tsx
@@ -1,0 +1,49 @@
+import styles from "./ReadMuse.module.css";
+import { MuseClient, zipSamples } from "muse-js";
+import { useState } from "react";
+
+export default function ReadMuse() {
+  const [isConnecting, setIsConnecting] = useState(false);
+  const [connected, setConnected] = useState(false);
+  const [client, setClient] = useState<MuseClient | null>(null);
+
+  return (
+    <div className={styles.page}>
+      <main className={styles.main}>
+        {!connected ? (
+          <button onClick={connect}>Connect</button>
+        ) : (
+          <button onClick={disconnect}>Disconnect</button>
+        )}
+      </main>
+    </div>
+  );
+
+  async function connect() {
+    try {
+      setIsConnecting(true);
+
+      let client = new MuseClient();
+      await client.connect();
+      await client.start();
+
+      client.eegReadings.subscribe((reading) => {
+        console.log(reading);
+      });
+
+      setClient(client);
+      setConnected(true);
+    } catch (err) {
+      console.error("connect():", err);
+      setClient(null);
+      setConnected(false);
+    } finally {
+      setIsConnecting(false);
+    }
+  }
+
+  async function disconnect() {
+    setClient(null);
+    setConnected(false);
+  }
+}
