+++ frontend/neuromance/src/app/components/Registration/Registration.tsx
@@ -1,0 +1,11 @@
+import styles from "./Registration.module.css";
+
+export default function Registration() {
+  return (
+    <div className={styles.page}>
+      <main className={styles.main}>
+        <h1>Registration</h1>
+      </main>
+    </div>
+  );
+}
